scripts/keycloak_kickstarter/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
ecdsa==0.18.0
idna==3.4
pyasn1==0.4.8
python-jose==3.3.0
python-keycloak==2.9.0
requests==2.28.2
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
urllib3==1.26.14
